Refactor settings and reformat multiple files
Previously, API URLs were hardcoded in the body of
manifester/manifester.py, and the offline token (which is used to
generate an access token, which is in turn used to authenticate API
calls) was a top-level setting. This commit makes the offline_token
setting a child setting of each manifest_category, and it moves the API
URLs to the settings file, also as children of each manifest_category.
This should enable greater flexibility by letting users associate
manifests with multiple RHSM accounts (potentially in multiple
environments, such as Red Hat's staging environment) in a single
settings file.

This commit also makes numerous formatting changes as a result of
running pre-commit.

Additionally, the offline_token validator in manifester/settings.py has
been commented out as it fails to validate with the offline_token
setting nested under the manifest categories.