Modify settings and requests to support proxies
Access to the internal Red Hat staging environment is mediated through a
proxy server. To support compatibility with this environment, this
commit adds a `proxies` setting to each manifest_category and adds a
`proxies` argument to each request in manifester.py. To access the
production RHSM API without a proxy, the value of the dictionary in the
`proxies` setting can be left as an empty string (i.e. `proxies:
{"https": ""}).