Add delete method to Manifester class
This commit adds a new method to the manifester class,
delete_subscription_allocation. This is necessary to prevent a buildup
of subscription allocations in RHSM accounts and to ensure that
subscriptions are not consumed by subscription allocations after those
allocations are no longer needed. Note that the first step in this
method is `self._access_token = None` because access tokens are only
valid for 5 minutes. Since it is likely that allocations will need to be
deleted more than five minutes after they are created, this step ensures
that a fresh, valid access token is passed with the DELETE request.