Refactor to accept dictionary as manifest category
After discussing this PR in the Satellite Automation channel, I am going
in a different direction here. Instead of ingesting the settings from
external frameworks, Manifester's `__init__` method will now accept a
either a dictionary or a string as the `manifest_category` argument.
This should allow greater flexibility with how Manifester's settings are
stored and accessed when used with external frameworks and also reduce
the risk of settings collisions resulting from Manifester loading all
the settings of such a framework.