Box manifest data, allow top-level URL setting
This commit ensures that, when manifest_category is passed as a dict, it
is wrapped in a DynaBox instance. This ensures consistency with the
behavior when settings are read from manifester_settings.yaml and should
allow DynaConf validation of manifest_category data passed from an
external framework. This commit also ensures that, if the token request
URL and allocations API endpoint URL are not present in manifest_category,
they will be read from the top level of manifester_settings.yaml.