Allow adding entitlements from unlimited pools
The `rhsatqe` RHSM account contains a pool of Red Hat Beta Access
subscriptions with an unlimited quantity. Because this quantity is
represented as `-1` in the response to an API request for the details of
that pool, Manifester previously failed to add subscriptions from that
pool to the allocations it was creating. This PR enables Manifester to
pull entitlements from subscription pools that contain an unlimited
quantity of entitlements.