Sanitize sat_version and handle sat_version errors
When creating a subscription allocation, the RHSM API only accepts a
string in the form of `sat-X.Y` for the `sat_version` parameter. This PR
adds a helper function that will remove the Z-stream version if it is
included in the value of the `sat_version` setting. It also adds error
handling for errors caused by an invalid `sat_version` during
subscription allocation creation.