Disable SCA after creating allocation
The RHSM API does not support disabling Simple Content Access (SCA) on
subscription allocations at creation time. This PR adds a check after
creating an allocation and, if SCA is disabled in the settings for that
allocation, performs another API call to disable SCA on that allocation.
The endpoint for disabling SCA requires that the `simpleContentAccess`
parameter be passed as JSON data rather than as a URL parameter, hence
the use of the `json` parameter where other requests in Manifester use
`params`.