Add manifest filename as attribute
In some scenarios, it will be useful to have the filename of the
manifest archive immediately accessible as an attribute of the object
returned by manifester. This commit assigns a manifest name based on the
subscription allocation name to the Manifester object during
instantiation and subsequently assigns that name to the response object
containing the manifest returned by the API's export job.