Fail over to the latest valid sat_version
With the adoption of Satellite stream, the logic for selecting a
sat_version for subscription allocations created by manifester is no
longer valid in all cases. This PR modifies the logic to default to the
most recent valid version if initial processing of the provided
sat_version does not yield a valid version.