Modify content and return value after timeout
This PR expicitly sets manifest.content to `None` and returns the
manifest object when a manifest export job times out. This change
facilitates the failover to cloned manifests in Robottelo. Previously,
failed export jobs triggered an exception in Robottelo when manifester
fixtures attempted to upload manifests because, in those cases, the
manifest object did not contain a `content` attribute.