Add test, modify mock method returns
This commit adds a negative test for the timeout on exporting a
manifest. It also modifies the return values of the mocked HTTP verb
methods to reuse the same RhsmApiStub object throughout each execution,
rather than returning a new RhsmApiStub object with each method call. It
also modifies the manifester class and helper methods to support this
change.