Switch to ruff for linting/formatting
This commit switches from using black for formatting and ruff for
linting to using ruff for both of these. It also implements a more
complete pyproject.toml and includes many small changes to get the
project into a passing state for ruff. Additionally, it removes setup.py
and setup.cfg as these are now redundant with pyproject.toml.