Use manifest_data dict for all tests
Since the manifester repo does not contain a fully populated config
file, this commit switches all of the tests to use the manifest_data
dictionary defined in the test_manifester.py module as the value of
manifest_category when instantiating a Manifester object. It also add
Python 3.12 to the list of classifiers in pyproject.toml.