Add username_prefix to settings and allocation name
This pull request is part of an effort to implement an inventory system
for subscription allocations created by Manifester. In situations where
multiple different users are creating allocations using the same login
on a shared RHSM account, the RHSM API does not provide a means to
differentiate which allocations were created by which users. To
facilitate such differentiation, this PR adds a `username_prefix`
setting that is used in the construction of the allocation name. This
should enable syncing the allocations created by individual users to
their local inventories while avoiding collisions with allocations
created by other users.