Add basic inventory file management
This PR adds helper functions to load a local inventory file and to
replace the contents of that file. It also adds calls to the update
inventory method several times in the course of a manifester run to try
to avoid stale data being stored persistently in the inventory file in
the event that a manifester run terminates unexpectedly.