Improve CLI functionality and CLI tests
This PR is another step in the manifester inventory/CLI improvement
effort. It adds functionality to and testing for the CLI with the goal
of making Manifester useful for interactive users in addition to its
current primary usecase in automated workflows.

Requires https://github.com/SatelliteQE/manifester/pull/29, so starting
this in draft status.