Add manifester delete command
This commit adds a 'manifester delete' command and makes various changes
to support its implementation. Notably, it adds a `minimal_init` option
to Manifester's `__init__` method to allow instantiating Manifester
objects with a subset of the attributes required for a full Manifester
run. This change does require some modifications to the structure of the
settings file, which I will push in a subsequent commit.

This commit also uses the `minimal_init` option to add a CLI option to
sync the inventory from RHSM, implements some changes suggested in
review, and makes several other small tweaks.