Add `scripts` section to pyproject.toml
When I converted Manifester to pyproject.toml, I did not include a
`project.scripts` section. As a result, an executable for the Manifester
CLI is not installed when the Manifester package is installed. This PR
adds a `project.scripts` section to pyproject.toml so the Manifester CLI
is easily accessible.