Add offline token argument to inventory command
This PR modifies the inventory CLI command to accept an offline_token
argument. This should resolve an issue with generating an access token
that I'm encountering with a CI pipeline. With this change, the job can
pass an environment variable containing the offline token to the
Manifester CLI, which mirrors how we access the same token in another CI
use case.