Use session-scoped inventory by default
Currently, Manifester syncs its inventory multiple times while
generating a manifest to ensure that the inventory data matches with the
corresponding data in RHSM. For some use cases, though, this behavior is
not desirable. Notably, in a CI scenario in which multiple workers share
the same `username_prefix` setting, syncing the inventory with RHSM
prevents those workers from running `manifester delete --all` at the end
of each pipeline, as this could delete allocations that are still being
used by other workers.

This PR addresses this issue by converting the default inventory
behavior to not sync from RHSM unless a sync is explicitly invoked.